Bug fix/get traveler (#28)
* v1.47.0

* v1.48.0

* v1.49.0

* v1.50.0

* v1.51.0

* v1.52.0

* v1.53.0

* v1.54.0

* v1.55.0

* v1.56.0

* v1.57.0

* v1.58.0

* v1.59.0

* update config

* v1.113.0

* v1.114.0

* v1.115.0

* v1.116.0

* v1.117.0

* v1.118.0

* v1.119.0

* v1.120.0

* v1.121.0

* v1.122.0

* v1.123.0

* v1.124.0

* v1.125.0

* bifrost tracking

---------

+++ bifrost_builder/src/utilities/getBifrostTravelerId.ts
@@ -10,7 +10,14 @@ export const getBifrostTravelerId = async (): Promise<{
     LOCAL_STORAGE_BIFROST_TRAVELER_ID_KEY
   );
 
-  if (maybeExistingBifrostTravelerId) {
+  console.log(
+    `maybeExistingBifrostTravelerId: ${maybeExistingBifrostTravelerId}`
+  );
+
+  if (
+    maybeExistingBifrostTravelerId &&
+    !["", "undefined"].includes(maybeExistingBifrostTravelerId)
+  ) {
     localStorage.removeItem(LOCAL_STORAGE_BIFROST_TRAVELER_ID_KEY);
     localStorage.setItem(
       LOCAL_STORAGE_BIFROST_TRAVELER_ID_KEY,
@@ -24,7 +31,8 @@ export const getBifrostTravelerId = async (): Promise<{
 
   try {
     const { bifrostTravelerId } = await getOrCreateBifrostTravelerId();
-    console.log(`Received bifrostTravelerId: ${bifrostTravelerId}`);
+    console.log(`bifrostTravelerId: ${bifrostTravelerId}`);
+
     localStorage.setItem(
       LOCAL_STORAGE_BIFROST_TRAVELER_ID_KEY,
       bifrostTravelerId
